1) Updating info.txt (binary builds and more information is now available at http://www.tittel.net/gridworldsim/) 2) Removing unnecessary commons-*.jar files and updating .classpath files for Eclipse
Two things are still not perfect:
1) If the Eclipse project for "commons" has locally a different name than "Tweety (GridWorldSim - Commons)", Eclipse will not find the commons project,
which is necessary to build server and serverapi. Is there a way to make Eclipse recognize that dependency in a way that is agnostic to the local project name?
2) Instead of the log4j-1.2.16.jar from the generalLibraries/ directory, each GridWorldSim Eclipse project has its own log4j-1.2.16.jar. Just adding the log4j-1.2.16.jar
from generalLibraries/ would break stuff for others, though, since absolute paths are used. Is there a way to fix that?

Netbeans doesn't suffer from both of the issues, but maybe it's just my lack of Eclipse skill.